requirements.txt
discord.py
aiohttp
python-dotenv

# Optional (only if uploads enabled)
dropbox
google-cloud-storage
